Move numpy emulation functions into a separate file
It's important to keep in mind that this numpyLua is not trying to port
numpy from python to Lua, but simply to facilitate the NNfS examples.

By looking at the implementation of the functions in the library,
readers can learn one of the many ways to implement such operation in
Lua; and are encouraged to write their own approach.